Changed Weighted from int to double for better precision

+++ Graphs/src/main/java/com/artcoffer/utilities/graph/Weighted.java
@@ -2,6 +2,6 @@ package com.artcoffer.utilities.graph;
 
 public interface Weighted {
 
-	public int getWeight();
+	public double getWeight();
 	
 }
